Move verify orchagent running function to common helpers file
